requirements.txt
PyQt5
discord.py
